Update AdB task traits and constants
